+++ frontend/types/canvas.ts
@@ -1,0 +1,5 @@
+export type Stroke = {
+	points: number[][];
+	color: string;
+	width: number;
+};
